backend/requirements.txt
fastapi>=0.115.0  # anyio 4.x 지원
uvicorn[standard]>=0.30.0  # 최신 버전으로 업그레이드
numpy<2.0.0  # NumPy 2.0 이하 버전 사용
qdrant-client>=1.5.0  # Qdrant Vector DB
anthropic>=0.51.0
httpx>=0.27.0,<0.28.0  # mcp-server-fetch 호환
pydantic>=2.7.0  # Pydantic 버전을 pydantic-settings와 호환되도록 조정
pydantic-settings==2.9.1  # Pydantic v2 설정 관리 패키지
openai>=1.10.0,<2.0.0  # Langchain 호환을 위한 OpenAI 버전 업그레이드
python-dotenv==1.0.0
boto3>=1.34.0  # AWS SDK for Python (Secrets Manager 등)
pytz>=2023.3  # 타임존 처리를 위한 패키지
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.0.1
tiktoken>=0.5.2,<1.0.0  # Langchain 호환을 위한 tiktoken 버전 업그레이드
tenacity==8.2.3
google-generativeai>=0.3.0 # Gemini API
cachetools>=4.2.0 # For in-memory caching
prometheus-client>=0.14.1 # Prometheus metrics client
pandas>=1.5.0
psutil>=5.8.0  # 시스템 리소스 모니터링을 위한 패키지
redis>=4.5.0  # Redis 캐싱을 위한 패키지
beautifulsoup4>=4.12.0  # HTML 파싱을 위한 패키지 (인라인 이미지 처리용)
anyio>=4.7.0  # 비동기 I/O 라이브러리 (mcp, sse-starlette 호환)
# Langchain 의존성 추가 (Phase 1: RunnableParallel 적용) - 호환 버전으로 업데이트
langchain>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-core>=0.3.63,<1.0.0  # langchain-anthropic 호환
langchain-openai>=0.2.0,<1.0.0
langchain-qdrant>=0.2.0,<1.0.0  # Qdrant 벡터스토어 통합
# 성능 모니터링을 위한 추가 패키지
# logging==0.4.9.6 - 제거됨, Python 내장 모듈과 충돌

# GPU 임베딩 최적화를 위한 패키지 (선택적 설치)
sentence-transformers>=2.2.0  # GPU 기반 임베딩 모델
torch>=1.13.0  # PyTorch (CUDA 지원)
transformers>=4.21.0  # Hugging Face transformers

# 멀티테넌트 보안을 위한 암호화 패키지
cryptography>=41.0.0  # 테넌트별 설정 암호화
psycopg2-binary>=2.9.0  # PostgreSQL 지원 (선택적)